requirements.txt
requests
pysocks
beautifulsoup4
lxml
